Revert renaming for error prone checks
