+++ tools/mc-java/src/test/java/io/spine/tools/mc/java/gradle/ErrorProneChecksExtensionTest.java
@@ -42,18 +42,18 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
 @DisplayName("ErrorProneChecksExtension should")
-class ModelChecksExtensionTest {
+class ErrorProneChecksExtensionTest {
 
     private Project project;
-    private ModelChecksExtension extension;
+    private ErrorProneChecksExtension extension;
 
     @BeforeEach
     void setUp(@TempDir Path tempDirPath) {
         File tempDir = tempDirPath.toFile();
         project = newProject(tempDir);
         ExtensionContainer extensions = project.getExtensions();
-        extension = extensions.create(ModelChecksPlugin.extensionName(),
-                                      ModelChecksExtension.class);
+        extension = extensions.create(ErrorProneChecksPlugin.extensionName(),
+                                      ErrorProneChecksExtension.class);
     }
 
     @Test
@@ -61,14 +61,14 @@ class ModelChecksExtensionTest {
     void return_use_validating_builder_severity() {
         final Severity expected = ERROR;
         extension.useValidatingBuilderSeverity = expected;
-        final Severity actual = ModelChecksExtension.getUseValidatingBuilderSeverity(project);
+        final Severity actual = ErrorProneChecksExtension.getUseValidatingBuilderSeverity(project);
         assertEquals(expected, actual);
     }
 
     @Test
     @DisplayName("return null severity if not set")
     void return_null_use_validating_builder_severity_if_not_set() {
-        final Severity severity = ModelChecksExtension.getUseValidatingBuilderSeverity(project);
+        final Severity severity = ErrorProneChecksExtension.getUseValidatingBuilderSeverity(project);
         assertNull(severity);
     }
 }
