+++ tools/mc-java/src/test/java/io/spine/tools/mc/java/gradle/ErrorProneChecksPluginTest.java
@@ -27,34 +27,38 @@
 package io.spine.tools.mc.java.gradle;
 
 import org.gradle.api.Project;
+import org.gradle.api.plugins.ExtensionContainer;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 
-import static io.spine.tools.mc.java.gradle.ModelChecksPlugin.extensionName;
+import static io.spine.tools.mc.java.gradle.given.ModelCompilerTestEnv.newProject;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 /**
- * The Error Prone Checks plugin extension.
+ * This test contains very basic scenarios of the plugin usage.
  *
- * <p>Allows configuring severity for all the Spine-custom Error Prone checks applied to the
- * project.
- *
- * @see Severity
+ * <p>For the tests of actual plugin functionality, see {@code io.spine.tools.check} test suites
+ * from this module.
  */
-@SuppressWarnings("PublicField" /* required for exposing the property in Gradle. */)
-public class ModelChecksExtension {
+@DisplayName("ErrorProneChecksPlugin should")
+class ErrorProneChecksPluginTest {
 
-    public Severity useValidatingBuilderSeverity;
-
-    /**
-     * Creates an instance of the extension in the passed project.
-     */
-    public static void createIn(Project project) {
-        project.getExtensions()
-                .create(extensionName(), ModelChecksExtension.class);
+    @Test
+    @DisplayName("create Spine check extension")
+    void create_spine_check_extension() {
+        Project project = newProject();
+        project.getPluginManager()
+               .apply(ErrorProneChecksPlugin.class);
+        ExtensionContainer extensions = project.getExtensions();
+        Object found = extensions.findByName(ErrorProneChecksPlugin.extensionName());
+        assertNotNull(found);
     }
 
-    public static Severity getUseValidatingBuilderSeverity(Project project) {
-        ModelChecksExtension extension = (ModelChecksExtension)
-                project.getExtensions()
-                       .getByName(extensionName());
-        return extension.useValidatingBuilderSeverity;
+    @Test
+    @DisplayName("apply to empty project")
+    void apply_to_empty_project_without_exceptions() {
+        Project project = newProject();
+        project.getPluginManager()
+               .apply(ErrorProneChecksPlugin.class);
     }
 }
